LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-calculating-min-max-content-expected.txt

foo fooo
x xx x
foo fooofoo fooofoo fooo
x xx x
foo fooofooo

PASS max-content
PASS min-content
